requirements.txt
aiohttp==3.4.4
backcall==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
Flask==1.1.2
future==0.18.2
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jupyter-client==6.1.3
jupyter-core==4.6.3
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
websockets==6.0
Werkzeug==1.0.1
wincertstore==0.2
yarl==1.3.0